requirements.txt
supervisely
google-api-python-client
pytube
ffmpeg-python